Remove v2 prefix/postfix, merge BaseWykopAPI into WykopAPI
