+++ tests/functional/test_api.py
@@ -15,10 +15,10 @@ class FileMock(object):
         return
 
 
-class TestWykopAPIv2(object):
+class TestWykopAPI(object):
 
     @responses.activate
-    def test_simple(self, wykop_api_v2):
+    def test_simple(self, wykop_api):
         rtype = 'rtype'
         body_dict = {
             'data': 'data'
@@ -27,8 +27,8 @@ class TestWykopAPIv2(object):
 
         named_params = 'appkey/123456app/format/json'
         url = '{protocol}://{domain}/{rtype}/{named_params}'.format(
-            protocol=wykop_api_v2._protocol,
-            domain=wykop_api_v2._domain,
+            protocol=wykop_api._protocol,
+            domain=wykop_api._domain,
             rtype=rtype,
             named_params=named_params,
         )
@@ -40,12 +40,12 @@ class TestWykopAPIv2(object):
             content_type='application/json',
         )
 
-        response = wykop_api_v2.request(rtype)
+        response = wykop_api.request(rtype)
 
         assert response == body_dict
 
     @responses.activate
-    def test_rmethod(self, wykop_api_v2):
+    def test_rmethod(self, wykop_api):
         rtype = 'rtype'
         rmethod = 'rmethod'
         body_dict = {
@@ -55,8 +55,8 @@ class TestWykopAPIv2(object):
 
         named_params = 'appkey/123456app/format/json'
         url = '{protocol}://{domain}/{rtype}/{rmethod}/{named_params}'.format(
-            protocol=wykop_api_v2._protocol,
-            domain=wykop_api_v2._domain,
+            protocol=wykop_api._protocol,
+            domain=wykop_api._domain,
             rtype=rtype,
             rmethod=rmethod,
             named_params=named_params,
@@ -69,12 +69,12 @@ class TestWykopAPIv2(object):
             content_type='application/json',
         )
 
-        response = wykop_api_v2.request(rtype, rmethod)
+        response = wykop_api.request(rtype, rmethod)
 
         assert response == body_dict
 
     @responses.activate
-    def test_named_params(self, wykop_api_v2):
+    def test_named_params(self, wykop_api):
         rtype = 'rtype'
         rmethod = 'rmethod'
         named_params_dict = {
@@ -87,8 +87,8 @@ class TestWykopAPIv2(object):
 
         named_params = 'appkey/123456app/format/json/page/2'
         url = '{protocol}://{domain}/{rtype}/{rmethod}/{named_params}'.format(
-            protocol=wykop_api_v2._protocol,
-            domain=wykop_api_v2._domain,
+            protocol=wykop_api._protocol,
+            domain=wykop_api._domain,
             rtype=rtype,
             rmethod=rmethod,
             named_params=named_params,
@@ -101,13 +101,13 @@ class TestWykopAPIv2(object):
             content_type='application/json',
         )
 
-        response = wykop_api_v2.request(
+        response = wykop_api.request(
             rtype, rmethod, named_params=named_params_dict)
 
         assert response == body_dict
 
     @responses.activate
-    def test_post_params(self, wykop_api_v2):
+    def test_post_params(self, wykop_api):
         rtype = 'rtype'
         rmethod = 'rmethod'
         post_param_name = 'post_param_name'
@@ -122,8 +122,8 @@ class TestWykopAPIv2(object):
 
         named_params = 'appkey/123456app/format/json'
         url = '{protocol}://{domain}/{rtype}/{rmethod}/{named_params}'.format(
-            protocol=wykop_api_v2._protocol,
-            domain=wykop_api_v2._domain,
+            protocol=wykop_api._protocol,
+            domain=wykop_api._domain,
             rtype=rtype,
             rmethod=rmethod,
             named_params=named_params,
@@ -136,13 +136,13 @@ class TestWykopAPIv2(object):
             content_type='application/json',
         )
 
-        response = wykop_api_v2.request(
+        response = wykop_api.request(
             rtype, rmethod, post_params=post_params)
 
         assert response == body_dict
 
     @responses.activate
-    def test_file_params(self, wykop_api_v2):
+    def test_file_params(self, wykop_api):
         rtype = 'rtype'
         rmethod = 'rmethod'
         file_param_name = 'file_param_name'
@@ -156,8 +156,8 @@ class TestWykopAPIv2(object):
 
         named_params = 'appkey/123456app/format/json'
         url = '{protocol}://{domain}/{rtype}/{rmethod}/{named_params}'.format(
-            protocol=wykop_api_v2._protocol,
-            domain=wykop_api_v2._domain,
+            protocol=wykop_api._protocol,
+            domain=wykop_api._domain,
             rtype=rtype,
             rmethod=rmethod,
             named_params=named_params,
@@ -170,13 +170,13 @@ class TestWykopAPIv2(object):
             content_type='application/json',
         )
 
-        response = wykop_api_v2.request(
+        response = wykop_api.request(
             rtype, rmethod, file_params=file_params)
 
         assert response == body_dict
 
     @responses.activate
-    def test_no_parser(self, wykop_api_v2):
+    def test_no_parser(self, wykop_api):
         rtype = 'rtype'
         rmethod = 'rmethod'
         body_dict = {
@@ -186,8 +186,8 @@ class TestWykopAPIv2(object):
 
         named_params = 'appkey/123456app/format/json'
         url = '{protocol}://{domain}/{rtype}/{rmethod}/{named_params}'.format(
-            protocol=wykop_api_v2._protocol,
-            domain=wykop_api_v2._domain,
+            protocol=wykop_api._protocol,
+            domain=wykop_api._domain,
             rtype=rtype,
             rmethod=rmethod,
             named_params=named_params,
@@ -200,13 +200,13 @@ class TestWykopAPIv2(object):
             content_type='application/json',
         )
 
-        response = wykop_api_v2.request(rtype, rmethod, parser=None)
+        response = wykop_api.request(rtype, rmethod, parser=None)
 
         assert response == body
 
     @mock.patch('wykop.api.requesters.urllib.urlopen')
     def test_urllib_requester(
-            self, mocked_urlopen, wykop_api_v2, urllib_requester):
+            self, mocked_urlopen, wykop_api, urllib_requester):
         rtype = 'rtype'
         rmethod = 'rmethod'
         body_dict = {
@@ -215,7 +215,7 @@ class TestWykopAPIv2(object):
         body = json.dumps(body_dict)
         mocked_urlopen.return_value = FileMock(body)
 
-        response = wykop_api_v2.request(
+        response = wykop_api.request(
             rtype, rmethod, requester=urllib_requester)
 
         assert mocked_urlopen.called
