--- tests/unit/api/v2/conftest.py
@@ -1,13 +1,0 @@
-import pytest
-
-from wykop.api.v2.clients import WykopAPIv2
-
-
-@pytest.fixture
-def wykop_api_v2():
-    return WykopAPIv2(
-        'sentinel.appkey',
-        'sentinel.secretkey',
-        output='sentinel.output',
-        response_format='sentinel.format',
-    )
