+++ wykop/__init__.py
@@ -1,5 +1,5 @@
 """Python library for the Wykop API."""
-from wykop.api.v2.clients import WykopAPIv2
+from wykop.api.client import WykopAPI
 from wykop.api.exceptions import WykopAPIError
 from wykop.utils import get_version
